pre-deployment fixes
