+++ src/assets/i18n/es.json
@@ -58,4 +58,5 @@
     "You have registered successfully!": "¡Te has registrado exitosamente!",
     "Please check your email and follow the steps to verify your account.": "Por favor revisa tu correo electrónico y sigue los pasos para verificar tu cuenta.",
     "Invalid email or password.": "Correo electrónico o contraseña inválidos."
+
 }
