+++ src/components/loginPages/accountRegistration/verifyAccWithCode.vue
