+++ src/components/loginPages/passReset/passResetRequest.vue
@@ -3,29 +3,23 @@
             <v-row>
                 <v-col cols="12" class="pb-0">
                     <h2 class="font-bold text-2xl text-red-700 tracking-widest">
-                        {{$t('Password Reset Code')}}
+                        {{$t('Password Reset')}}
                     </h2>
                 </v-col>
             </v-row>
             <v-row>
                 <v-col cols="12">
-                    {{$t('Please check your email and input the password reset code.')}}
-                    <br>
-                    {{$t('If you do not see an email, please check your Spam folder, or ')}}
-                    <span 
-                        class="font-semibold text-red-800 cursor-pointer"
-                        @click="$router.push('/testRegister')"
-                    >{{$t('Register for an Account.')}}</span>
+                    {{$t('Please enter your email. If it exists in our records, you will receive a confirmation code shortly to reset your password.')}}
                 </v-col>
             </v-row>
                 <v-row justify="center">
-                    <v-col cols="12" md="8">
+                    <v-col cols="12" md="10">
                         <v-sheet>
-                            <v-form ref="passForm" @submit.prevent="passFormSubmit">
+                            <v-form ref="emailForm" @submit.prevent="emailFormSubmit">
                                 <v-text-field
-                                    v-model="code"
+                                    v-model="email"
                                     :rules="rules"
-                                    :label="$t('Code:')"
+                                    :label="$t('Email:')"
                                     class="mx-auto"
                                     style="width: 100%;"
                                 >
@@ -59,80 +53,67 @@ import { useLoggedInUserStore } from "@/stored/loggedInUser";
 export default {
     data() {
         return {
-            userID: null,
+            userID: "",
             loading: false,
-            code: null,
+            email: null,
             rules: [
                 v => {
                     if (!v) {
-                        return this.$t('Code is required');
+                        return this.$t('Email is required');
+                    } else if (!/.+@.+/.test(v)) {
+                        return this.$t('Email must be valid');
                     }
                     return true;
                 }
             ],
         };
     },
-    mounted() {
-        if (this.$route.params && this.$route.params.userID) {
-            this.userID = this.$route.params.userID; 
-        }
-    },
     methods: {
-        async passFormSubmit() {
+        async emailFormSubmit() {
             // Check if there are any errors in the form
-            await this.$refs.passForm.validate();
-            const passFormInvalid = this.$refs.passForm.errors.length > 0;
+            await this.$refs.emailForm.validate();
+            const emailFormInvalid = this.$refs.emailForm.errors.length > 0;
 
-            // Instantiate the store
-            const loggedInUserStore = useLoggedInUserStore();
-
-            // If no errors, proceed
-            if (!passFormInvalid) {
+            // If no errors, proceed with login
+            if (!emailFormInvalid) {
                 this.loading = true;
-                
-                let requestData = {
-                    userID: this.userID,
-                    confirmationCode: this.code 
+                let user = {
+                    email: this.email,
+                    error: this.error,
                 };
-
-                let apiURL = import.meta.env.VITE_ROOT_API + `/userdata/resetPassWithCode`;
+                let apiURL = import.meta.env.VITE_ROOT_API + `/userdata/resetPassword`;
 
                 try {
-                    const response = await axios.post(apiURL, requestData);
+                    const response = await axios.put(apiURL, user);
 
                     if (response.status === 200) {
-                        // Set token as global header
-                        loggedInUserStore.setTokenHeader(response.data.token);
-                        
-                        this.$router.push({
-                            name: 'testResetPassNew',
-                            params: { 
-                                userID: this.userID,
-                                toastType: 'success',
-                                toastMessage: this.$t('Success! You may now reset your password.'),
-                                toastPosition: 'top-right',
-                                toastCSS: 'Toastify__toast--create'
-                            }
-                        });
-
+                        this.userID = response.data.userID;
+                        // Grab the token from the response
+                        this.token = response.data.token;
                     } else {
-                        toast.error(this.$t('Invalid code. Please try again.'), {
+                        toast.error(this.$t('An error occurred. Please try again.'), {
                             position: 'top-right',
                             toastClassName: 'Toastify__toast--delete'
                         });
                     }
                 } catch (error) {
-                    toast.error(this.$t('Invalid code. Please try again.'), {
+                    toast.error(this.$t('An error occurred. Please try again.'), {
                         position: 'top-right',
                         toastClassName: 'Toastify__toast--delete'
                     });
                 } finally {
                     this.loading = false;
+                    this.$router.push({
+                        name: 'passResetCode',
+                        params: { userID: this.userID }
+                    });
+
+
                 }
             }
         },
         goBackToLogin() {
-            this.$router.push({name: 'testLogin'});
+            this.$router.push({name: 'login'});
         },
     }
 }
